requirements.txt
paho-mqtt
influxdb
python-dotenv

